✨Datos de caracterización por encuentro
