+++ aplicaciones/procesador/fuente/procesadorCaracterizacion.ts
@@ -14,15 +14,17 @@ import type {
   DefinicionSimple,
   ElementoLista,
   EncuentroCaracterizacion,
+  EncuentroCaracterizacionConteo,
   Indicador,
   ListasCaracterizacion,
   ListasColectivos,
   LlavesCaracterizacion,
   LlavesColectivos,
 } from '@/tipos/compartidos';
-import type { Errata, FilaCaracterizacion, FilaColectivos } from './tipos';
+import type { Errata, FilaCaracterizacion, FilaCaracterizacionConteo, FilaColectivos } from './tipos';
+import { conector } from './ayudas';
 
-const encuentrosCaracterizacion: EncuentroCaracterizacion[] = [];
+const encuentrosCaracterizacion: EncuentroCaracterizacionConteo[] = [];
 const listas: ListasCaracterizacion = {
   sedes: [],
   tiposSede: [],
@@ -50,10 +52,14 @@ function procesarLista(llaveLista: LlavesCaracterizacion, valor: string) {
   return { nombre, slug };
 }
 
+function procesarEncuentro(id: number) {
+  const idEncuentro = id;
+}
+
 export default async (
   ruta: string,
   tabla: string
-): Promise<{ datos: EncuentroCaracterizacion[]; errata: Errata[] }> => {
+): Promise<{ datos: EncuentroCaracterizacionConteo[]; errata: Errata[] }> => {
   const flujo = await getXlsxStream({
     filePath: ruta,
     sheet: tabla,
@@ -66,71 +72,88 @@ export default async (
   return new Promise((resolver) => {
     flujo.on('data', async ({ raw }) => {
       const fila = raw.arr as FilaCaracterizacion;
+      const id = fila[0];
       const numero = limpiarTextoSimple(fila[1]);
+      let encuentro: EncuentroCaracterizacionConteo;
 
       if (numero) {
-        const encuentro: EncuentroCaracterizacion = { id: fila[0], numero: numero };
+        const existeEncuentro = encuentrosCaracterizacion.find((encuentro) => encuentro.id === id);
 
-        /** Sedes */
-        if (fila[3]) {
-          const valorSedes = limpiarTextoSimple(fila[3]);
-          const sedes: DefinicionSimple[] = [];
+        if (!existeEncuentro) {
+          encuentro = { id: fila[0], numero: numero, sedes: [], tiposSede: [], roles: [], cargos: [] };
+          encuentrosCaracterizacion.push(encuentro);
+        } else {
+          encuentro = existeEncuentro;
 
-          // Las sedes no se separaron con ; sino con y. Ejemplo: "Bogotá y Medellin"
-          if (valorSedes.includes(' y ')) {
-            // En este caso separar primero las sedes y luego procesarlas.
-            valorSedes.split(' y ').forEach((sede) => {
-              const { nombre, slug } = procesarLista('sedes', sede);
-              sedes.push({ nombre, slug });
+          /**
+           * Sedes: cantidad de participantes por sede por encuentro
+           */
+          if (fila[3]) {
+            const nombreSede = slugificar(limpiarTextoSimple(fila[3]));
+
+            const existeSede: { slug: string; conteo: number } | undefined = encuentro.sedes?.find((sede) => {
+              return sede.slug === nombreSede;
             });
-          } else {
-            const { nombre, slug } = procesarLista('sedes', valorSedes);
-            sedes.push({ nombre, slug });
+
+            if (!existeSede) {
+              encuentro.sedes?.push({ slug: nombreSede, conteo: 1 });
+            } else {
+              existeSede.conteo++;
+            }
           }
 
-          encuentro.sedes = sedes;
-        } else {
-          errata.push({ fila: numeroFila, error: `No tiene SEDE.` });
+          /**
+           * Tipos Sedes: cantidad de participantes por tipo de sede por encuentro
+           */
+          if (fila[4]) {
+            const tipoSede = slugificar(limpiarTextoSimple(fila[4]));
+
+            const existeTipo: { slug: string; conteo: number } | undefined = encuentro.tiposSede?.find((tipo) => {
+              return tipo.slug === tipoSede;
+            });
+
+            if (!existeTipo) {
+              encuentro.tiposSede?.push({ slug: tipoSede, conteo: 1 });
+            } else {
+              existeTipo.conteo++;
+            }
+          }
+
+          /**
+           * Roles: cantidad de participantes por rol por encuentro
+           */
+          if (fila[5]) {
+            const nombreRol = slugificar(limpiarTextoSimple(fila[5]));
+
+            const existeRol: { slug: string; conteo: number } | undefined = encuentro.roles?.find((rol) => {
+              return rol.slug === nombreRol;
+            });
+
+            if (!existeRol) {
+              encuentro.roles?.push({ slug: nombreRol, conteo: 1 });
+            } else {
+              existeRol.conteo++;
+            }
+          }
+
+          /**
+           * Cargos/Áreas: cantidad de participantes por cargo o área por encuentro
+           */
+          if (fila[6]) {
+            const nombreCargo = slugificar(limpiarTextoSimple(fila[6]));
+
+            const existeCargo: { slug: string; conteo: number } | undefined = encuentro.cargos?.find((cargo) => {
+              return cargo.slug === nombreCargo;
+            });
+
+            if (!existeCargo) {
+              encuentro.cargos?.push({ slug: nombreCargo, conteo: 1 });
+            } else {
+              existeCargo.conteo++;
+            }
+          }
         }
-
-        /** tipo de sede */
-        if (fila[4]) {
-          const valorTipoSede = limpiarTextoSimple(fila[4]);
-          const tiposSede: DefinicionSimple[] = [];
-          const { nombre, slug } = procesarLista('sedes', valorTipoSede);
-          tiposSede.push({ nombre, slug });
-
-          encuentro.tiposSede = tiposSede;
-        } else {
-          errata.push({ fila: numeroFila, error: `No tiene SEDE.` });
-        }
-
-        /** rol */
-        if (fila[5]) {
-          const valorRol = limpiarTextoSimple(fila[5]);
-          const roles: DefinicionSimple[] = [];
-          const { nombre, slug } = procesarLista('roles', valorRol);
-          roles.push({ nombre, slug });
-
-          encuentro.roles = roles;
-        } else {
-          errata.push({ fila: numeroFila, error: `No tiene ROL.` });
-        }
-
-        /** cargo */
-        if (fila[6]) {
-          const valorCargo = limpiarTextoSimple(fila[6]);
-          const cargos: DefinicionSimple[] = [];
-          const { nombre, slug } = procesarLista('cargos', valorCargo);
-          cargos.push({ nombre, slug });
-
-          encuentro.cargos = cargos;
-        } else {
-          errata.push({ fila: numeroFila, error: `No tiene CARGO.` });
-        }
-
-        encuentrosCaracterizacion.push(encuentro);
-        console.log(encuentro);
+        //   encuentrosCaracterizacion.push(encuentro);
       }
 
       numeroFila++;
@@ -138,19 +161,13 @@ export default async (
 
     flujo.on('close', () => {
       // Aquí ya terminó de leer toda la tabla
-      construirRelacionesEncuentros();
 
       for (const lista in listas) {
         ordenarListaObjetos(listas[lista as keyof ListasCaracterizacion], 'slug', true);
       }
 
-      /*   encuentrosCaracterizacion.sort((a, b) => {
-        if (a.titulo.slug < b.titulo.slug) return -1;
-        else if (a.titulo.slug > b.titulo.slug) return 1;
-        return 0;
-      }); */
-
-      guardarJSON(listas, 'listasEncuentros');
+      // guardar datos encuentros (no listas)
+      guardarJSON(encuentrosCaracterizacion, 'encuentros');
       resolver({ datos: encuentrosCaracterizacion, errata });
     });
 
@@ -158,83 +175,4 @@ export default async (
       throw new Error(JSON.stringify(error, null, 2));
     });
   });
-
-  function construirRelacionesEncuentros() {
-    // Estos campos son los que se usan para crear relaciones
-    const campos: LlavesCaracterizacion[] = ['sedes', 'tiposSede', 'roles', 'cargos'];
-
-    encuentrosCaracterizacion.forEach((encuentro) => {
-      // Pasar por cada campo sobre los que queremos construir relaciones
-      campos.forEach((campoRelacion) => {
-        const datosRelacion = encuentro[campoRelacion];
-
-        if (datosRelacion) {
-          const { id } = encuentro;
-
-          campos.forEach((campo) => {
-            // Agregar datos de cada campo en todos los otros, excepto en sí mismo.
-            if (campoRelacion !== campo) {
-              // Si no hay datos para llenar entonces podemos salir y continuar.
-              if (!encuentro[campo]) return;
-
-              const slugsCampoProyecto = aplanarDefinicionesASlugs(datosRelacion);
-
-              slugsCampoProyecto.forEach((slug) => {
-                const existe = listas[campoRelacion].find((obj) => obj.slug === slug);
-
-                if (existe) {
-                  if (!encuentro[campo]) return;
-
-                  let datos = encuentro[campo];
-
-                  const elementosDondeConectar = aplanarDefinicionesASlugs(datos);
-
-                  llenarRelacion(
-                    elementosDondeConectar,
-                    listas[campo],
-                    existe.id,
-                    campoRelacion,
-                    `${id}`,
-                    existe.nombre
-                  );
-                } else {
-                  console.log('Esto no puede pasar');
-                }
-              });
-            }
-            // }
-          });
-        }
-      });
-    });
-  }
-
-  function llenarRelacion(
-    elementosDondeConectar: string[],
-    elementoLista: ElementoLista[],
-    indice: string,
-    campoRelacion: LlavesCaracterizacion,
-    id: string,
-    nombre: string
-  ) {
-    elementosDondeConectar.forEach((elementoConector) => {
-      const elementoALlenar = elementoLista.find((obj) => obj.slug === elementoConector);
-
-      if (elementoALlenar) {
-        const existe = elementoALlenar.relaciones.find((obj) => obj.id === indice);
-
-        if (!existe) {
-          elementoALlenar.relaciones.push({ conteo: 1, id: indice, nombre: nombre, tipo: campoRelacion });
-        } else {
-          existe.conteo++;
-        }
-
-        if (!elementoALlenar.colectivos?.includes(id)) {
-          elementoALlenar.colectivos?.push(id);
-        }
-      }
-
-      // console.log('poner', campoRelacion, 'como relacion en lista', elementoConector);
-    });
-  }
 };
